# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ImpactInput {
  /** Estimated crowd size */
  crowdSize: number;
  /** Duration in minutes */
  duration: number;
  /** Whether road closure is required */
  roadClosed: boolean;
  /** Road sector or junction name */
  sector: string;
  /** Type of event */
  eventType?: string;
}

export interface ImpactResult {
  success: boolean;
  model: string;
  sector: string;
  impactScore: number;
  riskLevel: string;
  confidence: number;
  avgDelayMinutes: number;
  queueLengthVehicles: number;
  factors: string[];
}

export interface CongestionInput {
  crowdSize: number;
  duration: number;
  roadClosed: boolean;
}

export interface CongestionTimePoint {
  time: string;
  delay: number;
  queue: number;
  level: string;
}

export interface CongestionResult {
  success: boolean;
  sector: string;
  timeline: CongestionTimePoint[];
}

export interface ResourceInput {
  crowdSize: number;
  roadClosed: boolean;
  severity: string;
}

export interface ResourceAllocation {
  location: string;
  share: string;
  personnel: number;
}

export interface ResourceRequirements {
  police: number;
  marshals: number;
  barricades: number;
}

export interface ResourceResult {
  success: boolean;
  optimizationMethod: string;
  badge: string;
  requirements: ResourceRequirements;
  allocations: ResourceAllocation[];
}

export interface DiversionInput {
  affectedRoad: string;
}

export interface DiversionResult {
  success: boolean;
  affectedRoad: string;
  diversionType: string;
  normalEta: string;
  congestedEta: string;
  divertedEta: string;
  primaryRoute: string;
  secondaryRoute: string;
  savingMinutes: number;
  signageBoardsCount: number;
  status: string;
}

export interface EventAnalysisInput {
  eventType: string;
  location: string;
  crowdSize: number;
  roadClosed: boolean;
}

export interface EventAnalysis {
  strategicOverview: string;
  criticalAnomalies: string[];
  tacticalPlan: string[];
  signalOptimizations: string;
}

export interface EventAnalysisResult {
  success: boolean;
  source: string;
  analysis: EventAnalysis;
}

export interface TrafficEvent {
  id: string;
  eventType: string;
  location: string;
  status: string;
  priority: string;
  startTime: string;
  impactScore: number;
  /** @nullable */
  description?: string | null;
}

export interface EventStats {
  totalActive: number;
  criticalCount: number;
  avgImpactScore: number;
  resolvedToday: number;
  corridorsCovered: number;
  responseTimeMinutes: number;
}