# lib/api-zod/src/generated/types/congestionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CongestionInput {
  crowdSize: number;
  duration: number;
  roadClosed: boolean;
}