# lib/api-zod/src/generated/types/congestionTimePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CongestionTimePoint {
  time: string;
  delay: number;
  queue: number;
  level: string;
}