# lib/api-zod/src/generated/types/diversionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface DiversionResult {
  success: boolean;
  affectedRoad: string;
  diversionType: string;
  normalEta: string;
  congestedEta: string;
  divertedEta: string;
  primaryRoute: string;
  secondaryRoute: string;
  savingMinutes: number;
  signageBoardsCount: number;
  status: string;
}