# lib/api-zod/src/generated/types/eventAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { EventAnalysis } from './eventAnalysis';

export interface EventAnalysisResult {
  success: boolean;
  source: string;
  analysis: EventAnalysis;
}