# lib/api-zod/src/generated/types/eventStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface EventStats {
  totalActive: number;
  criticalCount: number;
  avgImpactScore: number;
  resolvedToday: number;
  corridorsCovered: number;
  responseTimeMinutes: number;
}