# lib/api-zod/src/generated/types/impactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ImpactInput {
  /** Estimated crowd size */
  crowdSize: number;
  /** Duration in minutes */
  duration: number;
  /** Whether road closure is required */
  roadClosed: boolean;
  /** Road sector or junction name */
  sector: string;
  /** Type of event */
  eventType?: string;
}