# lib/api-zod/src/generated/types/impactResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ImpactResult {
  success: boolean;
  model: string;
  sector: string;
  impactScore: number;
  riskLevel: string;
  confidence: number;
  avgDelayMinutes: number;
  queueLengthVehicles: number;
  factors: string[];
}