# lib/api-zod/src/generated/types/resourceAllocation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceAllocation {
  location: string;
  share: string;
  personnel: number;
}