# lib/api-zod/src/generated/types/resourceRequirements.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceRequirements {
  police: number;
  marshals: number;
  barricades: number;
}