# lib/api-zod/src/generated/types/resourceResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceAllocation } from './resourceAllocation';
import type { ResourceRequirements } from './resourceRequirements';

export interface ResourceResult {
  success: boolean;
  optimizationMethod: string;
  badge: string;
  requirements: ResourceRequirements;
  allocations: ResourceAllocation[];
}