# lib/api-zod/src/generated/types/trafficEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BTFI — Bengaluru Traffic Flow Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface TrafficEvent {
  id: string;
  eventType: string;
  location: string;
  status: string;
  priority: string;
  startTime: string;
  impactScore: number;
  /** @nullable */
  description?: string | null;
}